refactor(config/eslint): rename ESlint 'builder'
