+++ src/config/eslintrc-react.js
@@ -1,3 +1,3 @@
-const {buildConfig} = require('./helpers/eslint')
+const {buildConfig} = require('./helpers/build-eslint')
 
 module.exports = buildConfig({withReact: true})
