+++ src/config/helpers/build-eslint.js
